Added the ability to get a preorder traversal list.
The preorder traversal (data, left, right) is useful for hinting at the tree's internal structure without needing to get at Node<T>.
It always starts at the root, then walks the right and then left sub-trees. If the next value is less than the previous, it was a left sub-tree.
If the next value is more than the previous, it was a right sub-tree. If that next value is more than several previous values in the sequence, then you can backtrack all the way to the first value in the sequence.
If the next value is even greater than the first, then you know that's the beginning of the right sub-tree traversal.
It takes some visualization to get it right in your mind each time, especially with more values in the tree.