All the tests now pass.
Yeah, I know that TreePrinter tests still need some cleanup (the last two tests have terrible names and don't actually assert anything),
but now that it's all green it's a great time to commit.

Relying on the list order from post order traversals for asserting balance/rotation correctness was getting really tedious, so I added the TreePrinter
utility class to help me visualize the structure of the trees before and after balancing. I don't know if I'll be keeping it around for much longer, but
it stays for now...