Fixed issues related to deleting from an AvlTree.
Perhaps the last test I added is doing too much in one test, but it served to suss out several problems.
* I was still leaving stale parent pointers around after deleting a root node that had only one child.
* Deleting the root node that had two children no longer deletes the root node, but just replaces its data
with either its direct inorder predecessor or successor, depending on which of its subtrees was heavier.
The node containing the replacement data is deleted instead, since it only ever has zero or 1 subtrees.
* Checking if an empty tree was balanced threw an NPE since the root node was null.  Incidentally, empty trees are always balanced.

Well, all the tests pass again now.  Now it's time to start trimming some of the fat....